refactor(pagination): extract state into usePagination hook
